+++ changelog.md
@@ -7,3 +7,4 @@ vX.X.X
 - Added flashing caret to end of terminal text.
 - Added "Exit" item to "File" menu, closes #75.
 - Changed colour of open/close button so that it is green/red, and added play/stop icons, closes #76.
+- Added ability to send characters to COM port, closes #77.